Minor: mute compiler warnings during unittests building

+++ unit_tests/test-framework/engine_test_helper.cpp
@@ -442,7 +442,7 @@ bool EngineTestHelper::assertEventExistsAtEnginePhase(const char *msg, action_s 
 	//std::cout << "executor->size():              " << executor->size() << std::endl;
 	//std::cout << "expected_action.getCallback():  0x" << std::hex << reinterpret_cast<size_t>(action_expected.getCallback()) << "; name: " << action_expected.getCallbackName() << std::endl;
 
-	for (size_t i = 0; i < executor->size(); i++) {
+	for (int i = 0; i < executor->size(); i++) {
 		auto event = executor->getForUnitTest(i);
 		assert(event != nullptr);
